implement template Query class and derive Query classes from it
